+++ deprecated.go
@@ -90,3 +90,11 @@ var KeepFileMode = WithKeepFileMode
 // Deprecated: Use WithFileMode() instead.
 // This function may be removed in the future.
 var UseFileMode = WithFileMode
+
+// Deprecated: Use O_FILE_FLAG_DELETE_ON_CLOSE instead.
+// This constant may be removed in the future.
+const O_DELETE = O_FILE_FLAG_DELETE_ON_CLOSE
+
+// Deprecated: Use O_FILE_FLAG_NO_RO_ATTR instead.
+// This constant may be removed in the future.
+const O_NOROATTR = O_FILE_FLAG_NO_RO_ATTR