+++ walk.go
@@ -15,7 +15,7 @@ type FS = fs.FS
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// Source: https://github.com/golang/go/blob/77f911e3/src/io/fs/walk.go#L12-L128
+// Source: https://github.com/golang/go/blob/ac803b59/src/io/fs/walk.go#L12-L128
 
 // SkipDir is used as a return value from [WalkDirFunc] to indicate that
 // the directory named in the call is to be skipped. It is not returned