+++ CHANGELOG.md
@@ -15,6 +15,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Changed
 
+- Rename `O_DELETE` constant to `O_FILE_FLAG_DELETE_ON_CLOSE` to align with
+  upstream.
+- Rename `O_NOROATTR` constant to `O_FILE_FLAG_NO_RO_ATTR`.
+- Deprecate `O_DELETE` and `O_NOROATTR` constants.
+- Freshened golang code from upstream.
+- Always return `os.PathError` or `os.LinkError`, as appropriate.
+
 ## [v0.5.3](https://github.com/rasa/compat/compare/v0.5.2...v0.5.3)
 
 ### Added